#213 created card-status-widget
