+++ frontend/src/components/widgets/CardStatusChip.js
@@ -1,0 +1,48 @@
+import React from "react";
+import yellow from "@material-ui/core/colors/yellow";
+import orange from "@material-ui/core/colors/orange";
+import green from "@material-ui/core/colors/green";
+import red from "@material-ui/core/colors/red";
+import grey from "@material-ui/core/colors/grey";
+import blue from "@material-ui/core/colors/blue";
+import {
+  Paper,
+  Typography,
+} from "@material-ui/core";
+import { makeStyles } from "@material-ui/core/styles";
+
+import {
+  AGILE_CARD_STATUS_CHOICES,
+  BLOCKED,
+  READY,
+  IN_PROGRESS,
+  REVIEW_FEEDBACK,
+  IN_REVIEW,
+  COMPLETE,
+} from "../../constants";
+
+const useStyles = makeStyles((theme) => {
+  const card = {
+    margin: theme.spacing(1),
+    width: theme.spacing(15),
+    borderRadius: theme.spacing(2),
+    textAlign: "center",
+  };
+  return {
+    [BLOCKED]: { ...card, backgroundColor: grey[400] },
+    [READY]: { ...card, backgroundColor: blue[400] },
+    [IN_PROGRESS]: { ...card, backgroundColor: green[400] },
+    [REVIEW_FEEDBACK]: { ...card, backgroundColor: red[400] },
+    [IN_REVIEW]: { ...card, backgroundColor: orange[400] },
+    [COMPLETE]: { ...card, backgroundColor: yellow[400] },
+  };
+});
+
+export default ({ card }) => {
+  const classes = useStyles();
+  return (
+    <Paper className={classes[card.status]} variant="outlined">
+      <Typography>{AGILE_CARD_STATUS_CHOICES[card.status]}</Typography>
+    </Paper>
+  );
+};
