+++ frontend/src/stories/fixtures/repoProjectCard.json
@@ -2,7 +2,7 @@
   "id": 13095,
   "contentItem": 223,
   "contentItemUrl": "https://raw.githubusercontent.com/Umuzi-org/tech-department/master/content/projects/oop/person/_index.md",
-  "status": "IP",
+  "status": "B",
   "recruitProject": 2652,
   "assignees": [285],
   "reviewers": [72, 67, 294, 297],
